+++ tests/fixtures/project-yarn-sibling-node-modules-check/app/bower.json
